Import binutils-2.18-mint-20080119.diff
Fixed the linker script broken in the previous build.
The native ld wasn't able to find the standard libraries anymore.
Thanks to Miro "MiKRO" Kropáček for reporting this bug.

Fixed ld link failure when building the binutils on a system where
/bin/sh defines $BASH but does not support $BASH_LINENO
(this is the case on the current EasyMiNT).
Thanks to Keith Scroggins for reporting this bug.
https://sourceware.org/bugzilla/show_bug.cgi?id=5652

Included the patches needed to build the binutils natively on MiNT
with GCC 2.95.3.