Import binutils-2.18-mint-20080212.patch
The argument of the .align directive is now an absolute number
of bytes instead of a power of two. The behavior is now the same
as the .balign directive. It is consistent with the GAS documentation
about the m68k target, and with the ALIGN directive of the Pure Assembler.
Thanks to Roger Burrows for reporting the behavior of the native assemblers.