Import binutils-2.18-mint-20080228.patch
Restored the default section alignment to the default of 4 bytes.
Now the ".align 4" directive is honored as expected (with Falcon
TOS, or MiNT). The little drawback is that the size of every section
of every .o file is now rounded up to the next multiple of 4.
Thus some 2-byte fillers are inserted into the final executable
in order to make the actual alignment.

This reverts commit 7a50f87.