Import binutils-2.19.1-mint-20091004.patch
as: Fixed invalid generation of references to weak data symbols
defined in the same file. In practice, such construct was only
generated in C++ when using static data members in templates.
This was the cause of the "cout bug", which prevented to output
an int with cout, followed by some other text.
https://sourceware.org/bugzilla/show_bug.cgi?id=3041#c16

ld: Check for output sections VMA in a more compatible way.
This will help for future ELF support.