Import binutils-2.19.1-mint-20091017.patch
as: The behaviour of .align has been reverted to be compatible with
binutils 2.13 for MiNT. Now .align behaves like .p2align. So .align 2
is now equivalent to .p2align 2 and .balign 4.
You are strongly encouraged to use .balign (or .p2align) instead of
the unportable .align
This reverts commit 420fd774bb30bd80bc7902d33acdf078b63fad51

objdump: Customized -p for displaying the contents of the MiNT
extended header: stack size, start address...

strip: Executables linked with --traditional-format are now supported.

ld, strip: Executables with no symbols always declare an empty GNU
symbol table. So executables linked with -s or compiled with
--traditional-format then stripped are now byte-to-byte identical.

ld: Executables are now rejected as linker input.

objcopy: When the output is a MiNT executable, only a MiNT executable
is allowed as input.

Fixed compilation problem with -Wformat-security (enabled by default
on Ubuntu).
Thanks to David Gálvez for reporting this bug.

Big cleanup in prg-mint.c