Disable REGISTER_PREFIX_OPTIONAL
This is the default behaviour on standard ELF targets. It is necessary
to avoid clashes between register names and C variable/function names
when the leading underscore isn't used. So inside assembler sources, all
register names must be prefixed by the % character. And inside C
sources, in inline assembly blocks, the prefix must be escaped to %%.

It is still possible to avoid the % prefix requirement by adding
--register-prefix-optional to the as command line.